+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.0 under development
 
-- Chg #72: Change `Widget::widget()` method to using constructor arguments only or whole array definition (@vjik)
+- Chg #72, #80: Change `Widget::widget()` method to using constructor arguments only or whole array definition (@vjik)
 - Chg #74: Remove from abstract `Widget` class methods `beforeRun()`, `afterRun()` and `run()`, use `render()` only
   instead it (@vjik)